requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==4.2.11
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
requests==2.32.3
google-generativeai==0.8.4
edge-tts==6.1.15
PyPDF2==3.0.1
Pillow==11.1.0
gunicorn==25.3.0
packaging==26.1
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
cloudinary==1.42.1
django-cloudinary-storage==0.3.0
sib-api-v3-sdk==7.6.0
